feat: refactor search page + move files to correct folders in components
