+++ app/layout.tsx
@@ -4,8 +4,8 @@ import { Roboto } from "next/font/google";
 import Script from "next/script";
 import "./globals.css";
 import "react-tooltip/dist/react-tooltip.css";
-import Navbar from "@/components/Navbar";
-import Footer from "@/components/Footer";
+import Navbar from "@/components/layout/Navbar";
+import Footer from "@/components/layout/Footer";
 
 const roboto = Roboto({
   weight: ["100", "300", "400", "500", "700", "900"],
