+++ components/modals/ClanInfoModal.tsx
