+++ components/search/SearchHistory.tsx
