+++ components/search/SearchTabs.tsx
@@ -10,10 +10,10 @@ export default function SearchTabs({
   onTabChange,
 }: SearchTabsProps) {
   return (
-    <div className="flex p-1 mb-8 rounded-2xl bg-black/20 border-2 border-white/5 backdrop-blur-md">
+    <div className="flex p-1 mb-2 rounded-2xl bg-black/20 border-2 border-white/5 backdrop-blur-md">
       <button
         onClick={() => onTabChange("player")}
-        className={`flex-1 flex items-center justify-center gap-3 py-4 rounded-xl text-sm font-semibold transition-all duration-300 ${
+        className={`flex-1 flex items-center justify-center gap-2 py-3 rounded-lg text-sm font-semibold transition-all duration-300 ${
           activeTab === "player"
             ? "bg-emerald-600/90 text-white shadow-[0_0_20px_rgba(16,185,129,0.2)] border border-emerald-500/50"
             : "text-gray-400 hover:text-white hover:bg-white/5 border border-transparent"
@@ -28,7 +28,7 @@ export default function SearchTabs({
       </button>
       <button
         onClick={() => onTabChange("clan")}
-        className={`flex-1 flex items-center justify-center gap-3 py-4 rounded-xl text-sm font-semibold transition-all duration-300 ${
+        className={`flex-1 flex items-center justify-center gap-2 py-3 rounded-lg text-sm font-semibold transition-all duration-300 ${
           activeTab === "clan"
             ? "bg-emerald-600/90 text-white shadow-[0_0_20px_rgba(16,185,129,0.2)] border border-emerald-500/50"
             : "text-gray-400 hover:text-white hover:bg-white/5 border border-transparent"
